requirements.txt
adal==1.2.7
antlr4-python3-runtime==4.7.2
applicationinsights==0.11.10
argcomplete==1.12.3
azure-appconfiguration==1.1.1
azure-batch==10.0.0
azure-cli==2.24.2
azure-cli-core==2.24.2
azure-cli-telemetry==1.0.6
azure-common==1.1.27
azure-core==1.15.0
azure-cosmos==3.2.0
azure-datalake-store==0.0.52
azure-functions==1.7.2
azure-functions-devops-build==0.0.22
azure-graphrbac==0.60.0
azure-identity==1.6.0
azure-keyvault==1.1.0
azure-keyvault-administration==4.0.0b3
azure-loganalytics==0.1.0
azure-mgmt-advisor==2.0.1
azure-mgmt-apimanagement==0.2.0
azure-mgmt-appconfiguration==1.0.1
azure-mgmt-applicationinsights==0.1.1
azure-mgmt-authorization==0.61.0
azure-mgmt-batch==9.0.0
azure-mgmt-batchai==2.0.0
azure-mgmt-billing==1.0.0
azure-mgmt-botservice==0.3.0
azure-mgmt-cdn==11.0.0
azure-mgmt-cognitiveservices==6.3.0
azure-mgmt-compute==20.0.0
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==1.5.0
azure-mgmt-containerregistry==3.0.0rc17
azure-mgmt-containerservice==11.2.0
azure-mgmt-core==1.2.2
azure-mgmt-cosmosdb==6.3.0
azure-mgmt-databoxedge==0.2.0
azure-mgmt-datalake-analytics==0.2.1
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==4.1.0
azure-mgmt-deploymentmanager==0.2.0
azure-mgmt-devtestlabs==4.0.0
azure-mgmt-dns==8.0.0
azure-mgmt-eventgrid==3.0.0rc9
azure-mgmt-eventhub==4.1.0
azure-mgmt-hdinsight==2.2.0
azure-mgmt-imagebuilder==0.4.0
azure-mgmt-iotcentral==4.1.0
azure-mgmt-iothub==2.0.0
azure-mgmt-iothubprovisioningservices==0.2.0
azure-mgmt-keyvault==9.0.0
azure-mgmt-kusto==0.3.0
azure-mgmt-loganalytics==8.0.0
azure-mgmt-managedservices==1.0.0
azure-mgmt-managementgroups==0.2.0
azure-mgmt-maps==0.1.0
azure-mgmt-marketplaceordering==1.1.0
azure-mgmt-media==3.1.0
azure-mgmt-monitor==2.0.0
azure-mgmt-msi==0.2.0
azure-mgmt-netapp==2.0.0
azure-mgmt-network==19.0.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==0.5.0
azure-mgmt-privatedns==1.0.0
azure-mgmt-rdbms==8.1.0
azure-mgmt-recoveryservices==0.4.0
azure-mgmt-recoveryservicesbackup==0.11.0
azure-mgmt-redhatopenshift==0.1.0
azure-mgmt-redis==7.0.0rc2
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==0.6.0
azure-mgmt-resource==18.0.0
azure-mgmt-search==8.0.0
azure-mgmt-security==0.6.0
azure-mgmt-servicebus==6.0.0
azure-mgmt-servicefabric==0.5.0
azure-mgmt-servicefabricmanagedclusters==1.0.0
azure-mgmt-signalr==0.4.0
azure-mgmt-sql==0.26.0
azure-mgmt-sqlvirtualmachine==0.5.0
azure-mgmt-storage==18.0.0
azure-mgmt-synapse==0.6.0
azure-mgmt-trafficmanager==0.51.0
azure-mgmt-web==2.0.0
azure-multiapi-storage==0.6.2
azure-nspkg==3.0.2
azure-storage-common==1.4.2
azure-synapse-accesscontrol==0.5.0
azure-synapse-artifacts==0.6.0
azure-synapse-spark==0.2.0
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.3.2
fabric==2.6.0
humanfriendly==9.2
idna==3.2
invoke==1.5.0
isodate==0.6.0
javaproperties==0.5.1
Jinja2==3.0.1
jmespath==0.10.0
jsmin==2.2.2
jsondiff==1.2.0
knack==0.8.2
MarkupSafe==2.0.1
mock==4.0.3
msal==1.12.0
msal-extensions==0.3.0
msrest==0.6.21
msrestazure==0.6.4
oauthlib==3.1.1
packaging==20.9
paramiko==2.7.2
pathlib2==2.3.5
pkginfo==1.7.0
portalocker==1.7.1
psutil==5.8.0
pycparser==2.20
PyGithub==1.38
Pygments==2.9.0
PyJWT==1.7.1
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyreadline==2.1
python-dateutil==2.8.1
pytz==2019.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
scp==0.13.4
semver==2.13.0
six==1.16.0
sshtunnel==0.1.5
tabulate==0.8.9
urllib3==1.26.5
vsts==0.1.25
vsts-cd-manager==1.0.2
websocket-client==0.56.0
xmltodict==0.12.0
